big update for version 2
