+++ dev_note.md
@@ -1,0 +1,9 @@
+- ensemble.py
+    Trev to fix input_dir using pn
+
+- Entire load module need to be further edits
+    Trev
+
+- OutputRecorder
+    Don't defacult to_hdf5
+    Let user do it in a seperated function
