+++ src/pywrdrb/pre/__init__.py
@@ -16,6 +16,14 @@
 from abc import ABC, abstractmethod
 from .. import get_pn_object
 
+from .flows import *
+from .observations import *
+from .predict_diversions import *
+from .predict_inflows import *
+from .operational_constants import *
+
+
+
 class DataPreprocessor(ABC):
     def __init__(self):
         """
