--- src/pywrdrb/recorders/__init__.py
@@ -1,1 +1,0 @@
-from pywrdrb.recorders.output_recorder import OutputRecorder
